Remove the concept of region ids and begin removing location ids.
Begin moving to the concept of a string handle. This is essentially just the address of a static string. The intent is that we no longer
need to maintain a list of all known regions or locations. Instead, we just pass along the address of the static strings they are
comprised of. A sink will copy the string contents from the location any time it first sees a particular id into a new event (not yet added)
and the receiving end will build a table of string ids from those messages.

One goal of this work is to remove the need for Yardstick to retain much global allocated state. Instead, only connections need to
maintain state (their list of known string handles), which will remove the need for Yardstick to allocate memory before the application
has initialized the library (we will not allow connections until initialized). This also removes much of the state that must be kept or
calculated at the site of an event.